loader.js created lint
