+++ src/components/users/Users.jsx
@@ -1,18 +1,5 @@
 import { Link, useLoaderData } from "react-router-dom";
 
-export async function loader() {
-  try {
-    const response = await fetch(
-      "https://random-data-api.com/api/v2/users?size=10"
-    );
-    const data = await response.json();
-    console.log(data);
-    return data;
-  } catch (error) {
-    console.error("Error fecthing users:", error);
-    return [];
-  }
-}
 
 export const Users = () => {
   const users = useLoaderData();
